% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{check_for_updates}
\alias{check_for_updates}
\title{check_for_updates}
\usage{
check_for_updates(
  gh_pat = get_gh_pat(silent = TRUE),
  packages = get_available_packages(gh_pat = get_gh_pat(silent = TRUE), repository_name =
    "package_list", file_path = "packages.yaml"),
  return_package_list = FALSE
)
}
\arguments{
\item{gh_pat}{GitHub PAT to access the package list}

\item{return_package_list}{should the list of packages be returned?}
}
\value{
(optional) a list with outdated and up to date packages
}
\description{
Checks if there are updates for any of the factverse packages.
}
\examples{
\dontrun{
library(factverse)
check_for_updates()
}
}
