% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_package.R
\name{describe_package}
\alias{describe_package}
\title{describe_package}
\usage{
describe_package(
  package_name,
  description_file = c("README.md", "DESCRIPTION"),
  gh_pat = get_gh_pat(silent = TRUE),
  packages = get_available_packages(gh_pat = get_gh_pat(silent = TRUE), repository_name =
    "package_list", file_path = "packages.yaml")
)
}
\arguments{
\item{package_name}{name of the package}

\item{description_file}{name of the file that should be downloaded. By default,
factverse will first look for the README.md, and then for the DESCRIPTION.}

\item{gh_pat}{GitHub PAT to access the package list}

\item{packages}{list with the names of the packages for each repository that
are available}
}
\description{
Fetches the description of a package from GitHub.
}
\details{
When using factverse, it can sometimes be difficult to figure out what each
of the packages does. This function fetches a description of the package from
GitHub.
}
\examples{
\dontrun{
describe_package("testpackage")
}
}
