% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_load.R
\name{.onAttach}
\alias{.onAttach}
\title{.onAttach}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{not used}

\item{pkgname}{not used}
}
\value{
nothing
}
\description{
This function is executed every time library(factverse) is called.
It just looks fancy.
}
