% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factverse-package.R
\docType{package}
\name{factverse-package}
\alias{factverse}
\alias{factverse-package}
\title{Package Management with GitHub}
\description{
The factverse package provides a simple way to distribute
packages using private GitHub repositories.
}
\details{
The main functions are:
\itemize{
\item \code{\link[=get_available_packages]{get_available_packages()}}: Get a list of all packages
that you can install with factverse.
\item \code{\link[=describe_package]{describe_package()}}: Packages provided by factverse are typically
not public. \code{describe_package} provides a glimpse into what each package
does by downloading the README or DESCRIPTION.
\item \code{\link[=install_essentials]{install_essentials()}}: Install all packages that are
essential for your work.
\item \code{\link[=install_package]{install_package()}}: Install a specific package available
from factverse
\item \code{\link[=check_for_updates]{check_for_updates()}}: Check if there are newer versions available
for the packages you installed with factverse.
\item \code{\link[=update_packages]{update_packages()}}: Update all your packages that were
installed with factverse
}

Developers will find a detailed introduction at \url{https://github.com/Factworks/factverse}.
}
