% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{get_available_packages}
\alias{get_available_packages}
\title{get_available_packages}
\usage{
get_available_packages(
  gh_pat = get_gh_pat(silent = TRUE),
  repository_name = "package_list",
  file_path = "packages.yaml"
)
}
\arguments{
\item{gh_pat}{GitHub PAT to access the package list. This PAT should be found
automatically from a local file with secrets. If not, you can also pass a custom
PAT to the function. Create the PAT with new_PAT.}

\item{repository_name}{name of the repository that contains the package list}

\item{file_path}{path to the yaml file in the repository that contains the package list}
}
\value{
list with available packages
}
\description{
Retrieves the names of all available packages for factverse
}
\examples{
\dontrun{
library(factverse)
get_available_packages()
}
}
