% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_pat.R
\name{get_gh_pat}
\alias{get_gh_pat}
\title{get_gh_pat}
\usage{
get_gh_pat(path_to_gh_pat = get_pat_file_path(), silent = FALSE)
}
\arguments{
\item{path_to_gh_pat}{path to the file containing the GitHub PAT}

\item{silent}{should the function inform about the successful retrieval of the API key?}
}
\value{
list with PATs
}
\description{
Retrieves the GitHub PAT from the local secrets to access the private GitHub
repositories. The file may contain multiple GitHub PATs, in which case the
first valid one is returned (the first one where the expiry date is in the future).
}
\examples{
\dontrun{
library(factverse)
test_files <- system.file("extdata", package = "factverse")
test_yaml <- system.file("extdata",
                         "RSecret_Example.yaml",
                         package = "factverse")

factverse::get_gh_pat(test_yaml)
}
}
