% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_pat.R
\name{get_home_directory}
\alias{get_home_directory}
\title{get_home_directory}
\usage{
get_home_directory()
}
\value{
string with home directory path
}
\description{
Get the home directory irrespective of the operating system used.
}
\keyword{internal}
