% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_pat.R
\name{get_latest}
\alias{get_latest}
\title{get_latest}
\usage{
get_latest(files)
}
\arguments{
\item{files}{vector with the files that should be checked}
}
\value{
a single string with the name of the latest file
}
\description{
Extracts the latest file from a list of yaml files that follow the naming
pattern xyz-YYYY-MM-DD.yaml based on the date in the name.
}
\examples{
library(factverse)
factverse:::get_latest(c("Factverse_Secrets_2023-11-01.yaml",
                         "Factverse_Secrets_2024-11-01.yaml",
                         "Factverse_Secrets_2024-10-01.yaml",
                         "Factverse_Secrets_2024-1-1.yaml"))
}
\keyword{internal}
