% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{get_latest_package_sha}
\alias{get_latest_package_sha}
\title{get_latest_package_sha}
\usage{
get_latest_package_sha(
  gh_pat = get_gh_pat(silent = TRUE),
  repository_owner,
  repository_name
)
}
\arguments{
\item{gh_pat}{GitHub PAT to access the package list}

\item{repository_owner}{name of the owner of the repository on GitHub}

\item{repository_name}{name of the repository that contains the package list}
}
\description{
Retrieves the latest sha of an R package from the main branch.
}
\keyword{internal}
