% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{get_latest_package_version}
\alias{get_latest_package_version}
\title{get_latest_package_version}
\usage{
get_latest_package_version(
  gh_pat = get_gh_pat(silent = TRUE),
  repository_owner,
  repository_name,
  file_path = "DESCRIPTION"
)
}
\arguments{
\item{gh_pat}{GitHub PAT to access the package list}

\item{repository_owner}{name of the owner of the repository on GitHub}

\item{repository_name}{name of the repository that contains the package list}

\item{file_path}{path to the yaml file in the repository that contains the package list}
}
\description{
Retrieves the latest version of an R package from the package description.
Currently only supports versions of the form [0-9].[0-9].[0-9].
}
\keyword{internal}
