% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{get_package_sha}
\alias{get_package_sha}
\title{get_package_sha}
\usage{
get_package_sha(package_name)
}
\arguments{
\item{package_name}{name of the package}
}
\value{
the SHA of the package
}
\description{
Retrieves the sha of an installed R package.
}
\keyword{internal}
