% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_pat.R
\name{get_pat_file_path}
\alias{get_pat_file_path}
\title{get_pat_file_path}
\usage{
get_pat_file_path(
  folder_name = ".RSecrets",
  file_name_pattern =
    "Factverse_Secrets_\\\\d{4}\\\\-(0?[1-9]|1[012])\\\\-(0?[1-9]|[12][0-9]|3[01]).yaml"
)
}
\arguments{
\item{folder_name}{specify the name of the folder in which the secrets file is stored}

\item{file_name_pattern}{specify the pattern of the file in which the secrets are stored}
}
\value{
the path of the file with the secrets
}
\description{
Retrieves the path of the file that stores the secrets (access tokens) for
GitHub.
}
\details{
factverse looks in three locations for the file with the PAT:

1) in the current working directory
2) (only for windows) in AppData/Local/R
3) in the home directory

It will return the first match it finds.
}
\keyword{internal}
