% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_pat.R
\name{get_R_directory}
\alias{get_R_directory}
\title{get_R_directory}
\usage{
get_R_directory()
}
\value{
string with path to the R folder
}
\description{
Get the R directory where R packages are installed. By default, this
will look for the installation in AppData. If this is not found, it uses
the first result. If the directory is ~/AppData/Local/R/win-library/4.4,
this function will return AppData/Local/R/
}
\keyword{internal}
