% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_essentials}
\alias{install_essentials}
\title{install_essentials}
\usage{
install_essentials(
  gh_pat = get_gh_pat(silent = TRUE),
  packages = get_available_packages(gh_pat = get_gh_pat(silent = TRUE), repository_name =
    "package_list", file_path = "packages.yaml"),
  ...
)
}
\arguments{
\item{gh_pat}{GitHub PAT to access the package list}

\item{packages}{list with the names of the packages for each repository that
are available}

\item{...}{additional arguments passed to pak::pkg_install()}
}
\description{
Installs all essential packages (core packages) indicated in a package list.
}
\examples{
\dontrun{
library(factverse)
install_essentials()
}
}
