% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_package}
\alias{install_package}
\title{install_package}
\usage{
install_package(
  package_name,
  gh_pat = get_gh_pat(silent = TRUE),
  packages = get_available_packages(gh_pat = get_gh_pat(silent = TRUE), repository_name =
    "package_list", file_path = "packages.yaml"),
  ...
)
}
\arguments{
\item{package_name}{name of the package that should be installed}

\item{gh_pat}{GitHub PAT to access the package list}

\item{packages}{list with the names of the packages for each repository that
are available}

\item{...}{additional arguments passed to pak::pkg_install()}
}
\description{
Installs a package from a private GitHub repository.
}
\examples{
\dontrun{
library(factverse)
install_package("factworks")
}
}
