% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_pat.R
\name{new_PAT}
\alias{new_PAT}
\title{new_PAT}
\usage{
new_PAT(account_name, PAT)
}
\arguments{
\item{account_name}{name of the GitHub account}

\item{PAT}{PAT to access the GitHub account repositories}
}
\value{
returns an object of class GHPAT
}
\description{
new_PAT
}
\examples{
new_PAT(account_name = "MyAccount",
        PAT = "ksjg123")
}
