% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{print.AvailablePackages}
\alias{print.AvailablePackages}
\title{print.AvailablePackages}
\usage{
\method{print}{AvailablePackages}(x, ...)
}
\arguments{
\item{x}{output from get_available_packages}

\item{...}{not used}
}
\description{
Print the output from get_available_packages
}
\examples{
\dontrun{
library(factverse)
get_available_packages()
}
}
