% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{update_packages}
\alias{update_packages}
\title{update_packages}
\usage{
update_packages(
  gh_pat = get_gh_pat(silent = TRUE),
  packages = get_available_packages(gh_pat = get_gh_pat(silent = TRUE), repository_name =
    "package_list", file_path = "packages.yaml"),
  ...
)
}
\arguments{
\item{gh_pat}{GitHub PAT to access the package list}

\item{packages}{list with the names of the packages for each repository that
are available}

\item{...}{additional arguments passed to remotes::install_github}
}
\description{
Updates all installed packages with the latest version.
}
\examples{
\dontrun{
library(factworks)
update_packages()
}
}
