% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_github_pat.R
\name{get_pat_file_path}
\alias{get_pat_file_path}
\title{get_pat_file_path}
\usage{
get_pat_file_path(folder_name = ".RSecrets", file_name = "RSecrets.yaml")
}
\arguments{
\item{folder_name}{specify the name of the folder in which the secrets file is stored}

\item{file_name}{specify the name of the file in which the secrets are stored}
}
\value{
the path of the file with the secrets
}
\description{
Retrieves the path of the file that stores the secrets (access tokens) for
GitHub.
}
\keyword{internal}
